chore: rename amqp-manager lib
