+++ src/lib/amqp-manager/channel-manager.ts
