+++ src/lib/amqp-manager/config-manager.ts
