+++ src/lib/amqp-manager/connection-manager.ts
