+++ src/lib/amqp-manager/publisher-manager.ts
