+++ src/modules/room/room.handlers.ts
@@ -17,9 +17,9 @@ import { unbindAllSubscriptions } from '../subscription/subscription.service';
 import { KeyNamespace } from '@/types/state.types';
 import { permissionsGuard } from '../guards/guards.service';
 import { DsPermission } from '@/types/permissions.types';
-import AmqpManager from '@/lib/topic-exchange/amqp-manager';
+import AmqpManager from '@/lib/amqp-manager/amqp-manager';
 import { WebSocket } from 'uWebSockets.js';
-import ChannelManager from '@/lib/topic-exchange/channel-manager';
+import ChannelManager from '@/lib/amqp-manager/channel-manager';
 import { addRoomHistoryMessage } from '../history/history.service';
 import { enqueueWebhookEvent } from '../webhook/webhook.service';
 import { WebhookEvent } from '@/types/webhook.types';
