+++ src/modules/user/user.service.ts
@@ -4,7 +4,7 @@ import { Logger } from 'winston';
 import * as repository from './user.repository';
 import { WebSocket } from 'uWebSockets.js';
 import { Session } from '@/types/session.types';
-import ChannelManager from '@/lib/topic-exchange/channel-manager';
+import ChannelManager from '@/lib/amqp-manager/channel-manager';
 import { getNspClientId } from '@/util/helpers';
 
 function getUserSubscriptionKeyName(connectionId: string, nspClientId?: string): string {
