+++ test/lib/message-handler.test.ts
@@ -1,7 +1,7 @@
 import '../__mocks__/external/rabbitmq-client';
 import { mockApp } from '../__mocks__/external/uWebsockets';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import MessageHandler from '@/lib/topic-exchange/message-handler';
+import MessageHandler from '@/lib/amqp-manager/message-handler';
 import { getMockSession } from '@/modules/session/session.mock';
 
 vi.mock('uWebSockets.js');
